feat(ui): extract ResponseDetailPanel to separate file #371
Move ResponseDetailPanel from McpMessageLogPanel to its own file and improve content display with JsonLanguageField.
