+++ core/src/main/kotlin/cc/unitmesh/devti/mcp/ui/McpMessageLogPanel.kt
@@ -304,40 +304,7 @@ class McpMessageLogPanel : JPanel(BorderLayout()) {
         }
         
         private fun formatJsonValue(element: JsonElement): String {
-            // Convert JsonElement to a nicely formatted string
             return element.toString()
-        }
-    }
-
-    private class ResponseDetailPanel : JPanel(BorderLayout()) {
-        private val headerPanel = JPanel(FlowLayout(FlowLayout.LEFT)).apply {
-            border = JBUI.Borders.empty(10)
-            background = JBColor(0xF8F9FA, 0x2B2D30)
-        }
-        
-        private val durationLabel = JBLabel().apply {
-            font = font.deriveFont(Font.BOLD)
-        }
-        
-        private val contentTextArea = JTextArea().apply {
-            isEditable = false
-            wrapStyleWord = true
-            lineWrap = true
-            border = JBUI.Borders.empty(10)
-        }
-        
-        init {
-            headerPanel.add(JBLabel("Response Duration:"))
-            headerPanel.add(durationLabel)
-            
-            add(headerPanel, BorderLayout.NORTH)
-            add(JBScrollPane(contentTextArea), BorderLayout.CENTER)
-        }
-        
-        fun displayMessage(message: McpMessage) {
-            durationLabel.text = message.duration?.toString()?.plus(" ms") ?: "N/A"
-            contentTextArea.text = message.content
-            contentTextArea.caretPosition = 0
         }
     }
 }
